+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+- Added `client` option to `Recipe.launch` and `api.launch_experiment` for providing an existing Beaker client.
+
 ## [v3.4.1](https://github.com/allenai/beaker-gantry/releases/tag/v3.4.1) - 2026-01-12
 
 ### Fixed